+++ test/bernstein_tests.jl
@@ -8,7 +8,7 @@ import ContinuumArrays: apply, MulQuasiMatrix
     @test d*l isa BernsteinDerivative
     @test l' isa BernsteinDerivative
 
-    @test basis(l) == l.b
+    @test all(basis(l)[j](0.3) == l[0.3, j] for j in eachindex(l))
     @test nbasis(l) == 2
     @test eachindex(l) == 0:1
     @test order(l) == 2